introduce Decoration and Selection models
