+++ packages/slate-hyperscript/test/fixtures/cursor-across-blocks-and-inlines.js
@@ -136,7 +136,7 @@ export const output = {
     ],
   },
   selection: {
-    object: 'range',
+    object: 'selection',
     anchor: {
       object: 'point',
       key: '0',
@@ -150,7 +150,6 @@ export const output = {
       offset: 1,
     },
     isFocused: true,
-    isAtomic: false,
     marks: null,
   },
 }
