+++ packages/slate-hyperscript/test/fixtures/cursor-across-blocks-end.js
@@ -70,7 +70,7 @@ export const output = {
     ],
   },
   selection: {
-    object: 'range',
+    object: 'selection',
     anchor: {
       object: 'point',
       key: '0',
@@ -84,7 +84,6 @@ export const output = {
       offset: 3,
     },
     isFocused: true,
-    isAtomic: false,
     marks: null,
   },
 }
