+++ packages/slate-hyperscript/test/fixtures/cursor-across-multiple-blocks-middle.js
@@ -91,7 +91,7 @@ export const output = {
     ],
   },
   selection: {
-    object: 'range',
+    object: 'selection',
     anchor: {
       object: 'point',
       key: '0',
@@ -105,7 +105,6 @@ export const output = {
       offset: 1,
     },
     isFocused: true,
-    isAtomic: false,
     marks: null,
   },
 }
