+++ packages/slate-hyperscript/test/fixtures/cursor-block-end.js
@@ -47,7 +47,7 @@ export const output = {
     ],
   },
   selection: {
-    object: 'range',
+    object: 'selection',
     anchor: {
       object: 'point',
       key: '0',
@@ -61,7 +61,6 @@ export const output = {
       offset: 3,
     },
     isFocused: true,
-    isAtomic: false,
     marks: null,
   },
 }
