+++ packages/slate-hyperscript/test/fixtures/cursor-custom-block-middle.js
@@ -53,7 +53,7 @@ export const output = {
     ],
   },
   selection: {
-    object: 'range',
+    object: 'selection',
     anchor: {
       object: 'point',
       key: '0',
@@ -67,7 +67,6 @@ export const output = {
       offset: 1,
     },
     isFocused: true,
-    isAtomic: false,
     marks: null,
   },
 }
