+++ packages/slate-hyperscript/test/fixtures/cursor-inline.js
@@ -82,7 +82,7 @@ export const output = {
     ],
   },
   selection: {
-    object: 'range',
+    object: 'selection',
     anchor: {
       object: 'point',
       key: '0',
@@ -96,7 +96,6 @@ export const output = {
       offset: 1,
     },
     isFocused: true,
-    isAtomic: false,
     marks: null,
   },
 }
