+++ packages/slate-hyperscript/test/fixtures/cursor-mark.js
@@ -67,7 +67,7 @@ export const output = {
     ],
   },
   selection: {
-    object: 'range',
+    object: 'selection',
     anchor: {
       object: 'point',
       key: '0',
@@ -81,7 +81,6 @@ export const output = {
       offset: 4,
     },
     isFocused: true,
-    isAtomic: false,
     marks: null,
   },
 }
