+++ packages/slate-react/src/utils/find-range.js
@@ -1,5 +1,4 @@
 import getWindow from 'get-window'
-import isBackward from 'selection-is-backward'
 import { IS_IE, IS_EDGE } from 'slate-dev-environment'
 
 import findPoint from './find-point'
@@ -63,8 +62,6 @@ function findRange(native, value) {
   const range = document.createRange({
     anchor,
     focus,
-    isBackward: isCollapsed ? false : isBackward(native),
-    isFocused: true,
   })
 
   return range
