+++ packages/slate/src/changes/on-selection.js
@@ -3,7 +3,7 @@ import isEmpty from 'is-empty'
 import logger from 'slate-dev-logger'
 import pick from 'lodash/pick'
 
-import Range from '../models/range'
+import Selection from '../models/selection'
 
 const Changes = {}
 
@@ -12,7 +12,7 @@ Changes.blur = change => {
 }
 
 Changes.deselect = change => {
-  const range = Range.create()
+  const range = Selection.create()
   change.select(range)
 }
 
@@ -549,13 +549,13 @@ Changes.moveToStartOfText = change => {
 }
 
 Changes.select = (change, properties, options = {}) => {
-  properties = Range.createProperties(properties)
+  properties = Selection.createProperties(properties)
   const { snapshot = false } = options
   const { value } = change
   const { document, selection } = value
   const props = {}
   let next = selection.setProperties(properties)
-  next = document.resolveRange(next)
+  next = document.resolveSelection(next)
 
   // Re-compute the properties, to ensure that we get their normalized values.
   properties = pick(next, Object.keys(properties))
