+++ packages/slate/src/constants/model-types.js
@@ -7,6 +7,7 @@
 const MODEL_TYPES = {
   BLOCK: '@@__SLATE_BLOCK__@@',
   CHANGE: '@@__SLATE_CHANGE__@@',
+  DECORATION: '@@__SLATE_DECORATION__@@',
   DOCUMENT: '@@__SLATE_DOCUMENT__@@',
   HISTORY: '@@__SLATE_HISTORY__@@',
   INLINE: '@@__SLATE_INLINE__@@',
@@ -16,6 +17,7 @@ const MODEL_TYPES = {
   POINT: '@@__SLATE_POINT__@@',
   RANGE: '@@__SLATE_RANGE__@@',
   SCHEMA: '@@__SLATE_SCHEMA__@@',
+  SELECTION: '@@__SLATE_SELECTION__@@',
   STACK: '@@__SLATE_STACK__@@',
   TEXT: '@@__SLATE_TEXT__@@',
   VALUE: '@@__SLATE_VALUE__@@',
