+++ packages/slate/src/index.js
@@ -1,7 +1,11 @@
+import './interfaces/common'
+import './interfaces/node'
+import './interfaces/range'
 import Block from './models/block'
 import Change from './models/change'
 import Changes from './changes'
 import Data from './models/data'
+import Decoration from './models/decoration'
 import Document from './models/document'
 import History from './models/history'
 import Inline from './models/inline'
@@ -15,6 +19,7 @@ import PathUtils from './utils/path-utils'
 import Point from './models/point'
 import Range from './models/range'
 import Schema from './models/schema'
+import Selection from './models/selection'
 import Stack from './models/stack'
 import Text from './models/text'
 import TextUtils from './utils/text-utils'
@@ -33,6 +38,7 @@ export {
   Change,
   Changes,
   Data,
+  Decoration,
   Document,
   History,
   Inline,
@@ -48,6 +54,7 @@ export {
   resetKeyGenerator,
   resetMemoization,
   Schema,
+  Selection,
   setKeyGenerator,
   Stack,
   Text,
@@ -60,6 +67,7 @@ export default {
   Block,
   Changes,
   Data,
+  Decoration,
   Document,
   History,
   Inline,
@@ -75,6 +83,7 @@ export default {
   resetKeyGenerator,
   resetMemoization,
   Schema,
+  Selection,
   setKeyGenerator,
   Stack,
   Text,
