+++ packages/slate/src/models/document.js
@@ -1,9 +1,9 @@
 import isPlainObject from 'is-plain-object'
-import logger from 'slate-dev-logger'
 import { List, Map, Record } from 'immutable'
 
 import KeyUtils from '../utils/key-utils'
 import MODEL_TYPES, { isType } from '../constants/model-types'
+import Node from './node'
 
 /**
  * Default properties.
@@ -66,17 +66,11 @@ class Document extends Record(DEFAULTS) {
     const document = new Document({
       key,
       data: new Map(data),
-      nodes: Document.createChildren(nodes),
+      nodes: Node.createList(nodes),
     })
 
     return document
   }
-
-  /**
-   * Alias `fromJS`.
-   */
-
-  static fromJS = Document.fromJSON
 
   /**
    * Check if `any` is a `Document`.
@@ -95,45 +89,6 @@ class Document extends Record(DEFAULTS) {
 
   get object() {
     return 'document'
-  }
-
-  get kind() {
-    logger.deprecate(
-      'slate@0.32.0',
-      'The `kind` property of Slate objects has been renamed to `object`.'
-    )
-    return this.object
-  }
-
-  get isVoid() {
-    logger.deprecate(
-      '0.38.0',
-      'The `Node.isVoid` property is deprecated, please use the `Schema.isVoid()` checking method instead.'
-    )
-
-    return this.get('isVoid')
-  }
-
-  /**
-   * Check if the document is empty.
-   * Returns true if all it's children nodes are empty.
-   *
-   * @return {Boolean}
-   */
-
-  get isEmpty() {
-    logger.deprecate('0.38.0', 'The `Node.isEmpty` property is deprecated.')
-    return !this.nodes.some(child => !child.isEmpty)
-  }
-
-  /**
-   * Get the concatenated text of all the document's children.
-   *
-   * @return {String}
-   */
-
-  get text() {
-    return this.getText()
   }
 
   /**
@@ -155,14 +110,6 @@ class Document extends Record(DEFAULTS) {
     }
 
     return object
-  }
-
-  /**
-   * Alias `toJS`.
-   */
-
-  toJS(options) {
-    return this.toJSON(options)
   }
 }
