+++ packages/slate/test/serializers/raw/serialize/preserve-selection-and-keys.js
@@ -42,7 +42,7 @@ export const output = {
     ],
   },
   selection: {
-    object: 'range',
+    object: 'selection',
     anchor: {
       object: 'point',
       key: '0',
@@ -57,7 +57,6 @@ export const output = {
     },
     isFocused: false,
     marks: null,
-    isAtomic: false,
   },
 }
