+++ packages/slate/test/serializers/raw/serialize/preserve-selection.js
@@ -37,7 +37,7 @@ export const output = {
     ],
   },
   selection: {
-    object: 'range',
+    object: 'selection',
     anchor: {
       object: 'point',
       path: [0, 0],
@@ -50,7 +50,6 @@ export const output = {
     },
     isFocused: false,
     marks: null,
-    isAtomic: false,
   },
 }
